+++ Game.Core/Domain/Actors/Actor.cs
@@ -3,23 +3,17 @@ namespace Game.Core.Domain.Actors;
 public abstract class Actor
 {
     public string Id { get; }
-    public ActorBaseStats BaseStats { get; }
-    public ActorStats Stats { get; private set; }
-    public int Level { get; }
+
+    public ActorStats Stats { get; protected set; }
+    public int Level { get; protected set; }
     public bool IsAlive => Stats.CurrentHp > 0;
 
-    protected Actor(string id, ActorBaseStats baseStats, int? level = null)
+    protected Actor(string id, ActorStats stats, int? level = null)
     {
         Id = id;
-        BaseStats = baseStats;
         Level = level ?? 1;
-        Stats = new(
-            currentHp: BaseStats.MaxHp * Level,
-            attack: BaseStats.Attack * Level,
-            defense: BaseStats.Defense * Level,
-            criticalChance: BaseStats.CriticalChance,
-            luck: BaseStats.Luck
-        );
+        Stats = stats;
+        Stats = RecalculateStats();
     }
 
     public void ReceiveDamage(int value)
@@ -29,6 +23,23 @@ public abstract class Actor
 
     public virtual void Heal(int value)
     {
-        Stats.CurrentHp = Math.Min(Stats.CurrentHp, Stats.CurrentHp + value);
+        Stats.CurrentHp = Math.Min(Stats.MaxHp, Stats.CurrentHp + value);
+    }
+
+    protected virtual ActorStats RecalculateStats()
+    {
+        return new(
+            maxHp: Stats.MaxHp + (Level * Stats.MaxHp),
+            attack: Stats.Attack + (Level * Stats.Attack),
+            defense: Stats.Defense + (Level * Stats.Defense),
+            criticalChance: Stats.CriticalChance,
+            luck: Stats.Luck
+        );
+    }
+
+    protected virtual void LevelUp()
+    {
+        Level += 1;
+        Stats = RecalculateStats();
     }
 }