+++ Game.Core/Domain/Actors/ActorStats.cs
@@ -2,15 +2,17 @@ namespace Game.Core.Domain.Actors;
 
 public class ActorStats
 {
+    public int MaxHp { get; set; }
     public int CurrentHp { get; set; }
     public int Attack { get; set; }
     public int Defense { get; set; }
     public int CriticalChance { get; set; }
     public int Luck { get; set; }
 
-    public ActorStats(int currentHp, int attack, int defense, int criticalChance, int luck)
+    public ActorStats(int maxHp, int attack, int defense, int criticalChance, int luck)
     {
-        CurrentHp = currentHp;
+        MaxHp = maxHp;
+        CurrentHp = maxHp;
         Attack = attack;
         Defense = defense;
         CriticalChance = criticalChance;