+++ Game.Core/Domain/Actors/Player.cs
@@ -2,15 +2,17 @@ namespace Game.Core.Domain.Actors;
 
 public class Player : Actor
 {
+    public PlayerType Type { get; private set; }
     public int Exp { get; private set; }
     public int ExpNextLevel { get; private set; }
     public int Gold { get; private set; }
 
-    public Player(string id, ActorBaseStats baseStats)
-        : base(id, baseStats)
+    public Player(string id, ActorStats stats, PlayerType type)
+        : base(id, stats)
     {
+        Type = type;
         Exp = 0;
-        ExpNextLevel = 100;
+        ExpNextLevel = CalculateExpNextLevel();
         Gold = 100;
     }
 
@@ -20,6 +22,13 @@ public class Player : Actor
             return;
 
         Exp += amount;
+
+        while (Exp >= ExpNextLevel)
+        {
+            Exp -= ExpNextLevel;
+            LevelUp();
+            ExpNextLevel = CalculateExpNextLevel();
+        }
     }
 
     public void AddGold(int amount)
@@ -28,5 +37,23 @@ public class Player : Actor
             return;
 
         Gold += amount;
+    }
+
+    private int CalculateExpNextLevel()
+    {
+        return (int)Math.Floor(100 * Math.Pow(Level, 1.5));
+    }
+
+    protected override ActorStats RecalculateStats()
+    {
+        var map = PlayerProgressionMap.Values;
+
+        return new(
+            maxHp: Stats.MaxHp + (Level * Stats.MaxHp) + map[Type].MaxHp,
+            attack: Stats.Attack + (Level * Stats.Attack) + map[Type].Attack,
+            defense: Stats.Defense + (Level * Stats.Defense) + map[Type].Defense,
+            criticalChance: Stats.CriticalChance,
+            luck: Stats.Luck
+        );
     }
 }