+++ Game.Core/Domain/Actors/PlayerType.cs
@@ -1,0 +1,8 @@
+namespace Game.Core.Domain.Actors;
+
+public enum PlayerType
+{
+    Warrior,
+    Mage,
+    Archer,
+}