+++ Game.Tests/Globals.cs
@@ -1,1 +1,9 @@
+global using Game.Core.Application.Combat.DTO;
+global using Game.Core.Application.Combat.UseCases;
+global using Game.Core.Application.Results;
+global using Game.Core.Domain.Actors;
+global using Game.Core.Domain.Combat;
+global using Game.Core.Domain.Combat.Actions;
+global using Game.Core.Domain.Combat.Reward;
+global using Game.Core.Extensions;
 global using Xunit;